[WIP] Style and small changes
Hi! This was just the library I was looking for. I'm hoping we can move away from hard coding the the dates positions so it'd be possible to dynamically change the width and height of the whole calendar. I know https://github.com/schmich/marinara has an open implementation that I could check out later. Thoughts?

Thanks!